Make things configurable
There are now some very simple rules available, shown in conf.yml. A
function (dim with level 100 is the only one supported) can have a
number of triggers. The only trigger rule available is until, which
fires the trigger until that number of seconds has passed since the
function in trigger.name fired.